+++ nemo/core/neural_types/neural_type.py
@@ -329,7 +329,6 @@ class NmTensor(NeuralType):
         """
         AppState().tensor_names.rename_NmTensor(self, new_name)
 
-
 class NeuralTypeError(Exception):
     """Base class for neural type related exceptions."""